Replaced math function with openlibmath

+++ library/math/copysignf.c
@@ -49,7 +49,7 @@
 float 
 copysignf(float x, float y)
 {
-	unsigned long ix, iy;
+    uint32_t ix, iy;
 	GET_FLOAT_WORD(ix, x);
 	GET_FLOAT_WORD(iy, y);
 	SET_FLOAT_WORD(x, (ix & 0x7fffffff) | (iy & 0x80000000U));
